client/requirements.txt
pynput==1.7.6
requests==2.31.0
pystray==0.19.5
Pillow==10.2.0
cx_Freeze==6.15.12
python_version >= "3.8"
winreg==0.0.0 